+++ src/Reader.php
@@ -3,6 +3,8 @@ declare(strict_types = 1);
 
 namespace Middlewares;
 
+use Middlewares\Utils\Factory;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\MiddlewareInterface;
@@ -17,12 +19,26 @@ class Reader extends Filesystem implements MiddlewareInterface
     private $continueOnError = false;
 
     /**
+     * @var ResponseFactoryInterface
+     */
+    private $responseFactory;
+
+    /**
      * Configure if continue to the next middleware if the file is not found.
      */
     public function continueOnError(bool $continueOnError = true): self
     {
         $this->continueOnError = $continueOnError;
 
+        return $this;
+    }
+
+    /**
+     * Set the response factory used.
+     */
+    public function responseFactory(ResponseFactoryInterface $responseFactory): self
+    {
+        $this->responseFactory = $responseFactory;
         return $this;
     }
 
@@ -37,7 +53,9 @@ class Reader extends Filesystem implements MiddlewareInterface
                 return $handler->handle($request);
             }
 
-            return Utils\Factory::createResponse(405)->withHeader('Allow', 'GET');
+            $responseFactory = $this->responseFactory ?: Factory::getResponseFactory();
+
+            return $responseFactory->createResponse(405)->withHeader('Allow', 'GET');
         }
 
         $file = static::getFilename($request->getUri()->getPath());
@@ -54,7 +72,9 @@ class Reader extends Filesystem implements MiddlewareInterface
                 return $handler->handle($request);
             }
 
-            return Utils\Factory::createResponse(404);
+            $responseFactory = $this->responseFactory ?: Factory::getResponseFactory();
+
+            return $responseFactory->createResponse(404);
         }
 
         return $this->read($request, $file)->withHeader('Content-Encoding', 'gzip');
@@ -71,7 +91,10 @@ class Reader extends Filesystem implements MiddlewareInterface
             throw new RuntimeException(sprintf('Unable to read the file "%s"', $file)); //@codeCoverageIgnore
         }
 
-        $response = Utils\Factory::createResponse()->withBody(Utils\Factory::createStream($resource));
+        $responseFactory = $this->responseFactory ?: Factory::getResponseFactory();
+        $streamFactory = $this->streamFactory ?: Factory::getStreamFactory();
+
+        $response = $responseFactory->createResponse()->withBody($streamFactory->createStreamFromResource($resource));
 
         return self::range($response, $request->getHeaderLine('Range'));
     }