+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - UNRELEASED
+
+### Added
+
+- New option `streamFactory` to configure a custom factory implementing `Psr\Http\Message\StreamFactoryInterface`
+- New option `responseFactory` (only in `Reader`) to configure a custom factory implementing `Psr\Http\Message\ResponseFactoryInterface`
+
+### Changed
+
+- Upgrade dev dependencies
+
 ## [1.0.0] - 2017-01-25
 
 ### Added
@@ -65,6 +76,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+
+[1.1.0]: https://github.com/middlewares/filesystem/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/middlewares/filesystem/compare/v0.5.0...v1.0.0
 [0.5.0]: https://github.com/middlewares/filesystem/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/middlewares/filesystem/compare/v0.3.1...v0.4.0